server/requirements.txt
flask
python-dotenv
flask-queue-sse
Flask-RESTx